+++ libs/cardano-protocol/src/Cardano/Protocol/TPraos/BlockHeader.hs
@@ -67,8 +67,10 @@ import Cardano.Ledger.Binary (
   EncCBOR (..),
   EncCBORGroup (..),
   TokenType (TypeNull),
+  decodeFixedSized,
   decodeNull,
   decodeRecordNamed,
+  encodeFixedSized,
   encodeListLen,
   encodeNull,
   listLenInt,
@@ -76,7 +78,6 @@ import Cardano.Ledger.Binary (
   runByteBuilder,
   serialize',
  )
-import Cardano.Ledger.Binary.Crypto
 import qualified Cardano.Ledger.Binary.Plain as Plain
 import Cardano.Ledger.Block (Block (..), EraBlockHeader (..))
 import Cardano.Ledger.Core (Era)
@@ -178,7 +179,7 @@ instance Crypto c => EncCBOR (BHBody c) where
       <> encCBOR (bheaderSlotNo bhBody)
       <> encCBOR (bheaderPrev bhBody)
       <> encCBOR (bheaderVk bhBody)
-      <> encodeVerKeyVRF (bheaderVrfVk bhBody)
+      <> encodeFixedSized (bheaderVrfVk bhBody)
       <> encCBOR (bheaderEta bhBody)
       <> encCBOR (bheaderL bhBody)
       <> encCBOR (bsize bhBody)
@@ -195,7 +196,7 @@ instance Crypto c => DecCBOR (BHBody c) where
     bheaderSlotNo <- decCBOR
     bheaderPrev <- decCBOR
     bheaderVk <- decCBOR
-    bheaderVrfVk <- decodeVerKeyVRF
+    bheaderVrfVk <- decodeFixedSized
     bheaderEta <- decCBOR
     bheaderL <- decCBOR
     bsize <- decCBOR
@@ -232,12 +233,12 @@ instance Crypto c => EncCBOR (BHeaderRaw c) where
     let BHeaderRaw {..} = bh
      in encodeListLen 2
           <> encCBOR bhrBody
-          <> encodeSignedKES bhrSignature
+          <> encodeFixedSized bhrSignature
 
 instance Crypto c => DecCBOR (BHeaderRaw c) where
   decCBOR =
     decodeRecordNamed "BHeaderRaw" (const 2) $
-      BHeaderRaw <$> decCBOR <*> decodeSignedKES
+      BHeaderRaw <$> decCBOR <*> decodeFixedSized
 
 instance Crypto c => DecCBOR (Annotator (BHeaderRaw c)) where
   decCBOR = pure <$> decCBOR