+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 1.21.0.0
 
 * Restrict `getProducedValue` to `TxBody TopTx era`.
+* Remove `decodeSignedDSIGNED`, `encodeSignedDSIGN` in favor of `encodeFixedSized`/`decodeFixedSized`
+* Remove `decodeSignedDSIGN`, `encodeSignedDSIGN` in favor of `encodeFixedSized`/`decodeFixedSized`
+* Remove `FromCBOR` and `ToCBOR` instances for `VKey`
 * Remove `Semigroup` and `Monoid` instances for `CostModels`
 * Change `updateCostModels` to take `CostModelsUpdate` as its second argument
 * Add `CostModelsUpdate`